+++ CHANGES.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+
+Version 5.0.1 (2021-06-xx)
+==========================
+* Add :meth:`~kartothek.io.eager.copy_dataset` to copy and optionally rename datasets within one store or between stores (eager only)
+* Add renaming option to :meth:`~kartothek.io.eager_cube.copy_cube`
+
+
 Version 5.0.0 (2021-06-23)
 ==========================